+++ src/search/index.test.ts
@@ -6,6 +6,8 @@ import IDBKeyRange from 'fake-indexeddb/lib/FDBKeyRange'
 
 import * as index from './'
 import * as oldIndex from './search-index-old'
+import { exportPages as exportOldPages } from './search-index-old/export'
+import { importPage as importNewPage } from './search-index-new/import'
 import * as newIndex from './search-index-new'
 import * as DATA from './index.test.data'
